% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcipubmedtool.R
\name{Multiple_search_validation_message}
\alias{Multiple_search_validation_message}
\title{多重搜尋驗證}
\usage{
Multiple_search_validation_message(path, diseases, SNP, race, PMIDs_manual)
}
\arguments{
\item{path}{想要輸出到哪個路徑資料夾儲存}

\item{SNP}{第二組關鍵字}

\item{race}{第三組關鍵字}

\item{PMIDs_manual}{將先前手動搜尋得到的PMIDs拿來驗證}

\item{disease}{第一關鍵字}
}
\value{
txt file
}
\description{
多重搜尋驗證
}
