% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcipubmedtool.R
\name{Processing_text_mining_from_entire_table}
\alias{Processing_text_mining_from_entire_table}
\title{針對已經建立好的rsID對應疾病的G2清單，共384組合進行表格輸出}
\usage{
Processing_text_mining_from_entire_table(
  numbers,
  data = G2_rsID_Disease_PMIDs_list_clean,
  pmid_list,
  file_names
)
}
\arguments{
\item{numbers}{匯入表格的row numbers}

\item{data}{完整的表格，包含疾病和rsID對應組合}

\item{pmid_list}{}

\item{file_names}{輸出的檔案名稱}
}
\value{
輸出大表格
}
\description{
針對已經建立好的rsID對應疾病的G2清單，共384組合進行表格輸出
}
