% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcipubmedtool.R
\name{Pubmed_ID_retrieve}
\alias{Pubmed_ID_retrieve}
\title{pubmed retrieve PMIDs 通用版本，可自定義多種關鍵字
此版本從先前自己寫的SNPs_pubmed_for_multiple_factors啟發和延伸，目的是為變成通用版
有時候get_pubmed_ids會出現warning: error in erl: 無法開啟...連結
20230215 已新增 return retrieve PMIDs
20230215 已解決 "ERBB2 AND Lung cancer AND drug" >> Note that only 0 PubMed IDs were retrieved (1170 were expected).}
\usage{
Pubmed_ID_retrieve(keyword_col_list, data)
}
\arguments{
\item{keyword_col_list}{檔案表格關鍵字的欄名，也是關鍵字的類別，例如: 生物標誌、癌症、藥物}

\item{data}{完整的表格，包含不同類型關鍵字，以欄(column)分別呈現，例如: EGFR, Lung cancer, herceptin}
}
\value{
輸出不同關鍵字排列組合對應的PMID列表
}
\description{
pubmed retrieve PMIDs 通用版本，可自定義多種關鍵字
此版本從先前自己寫的SNPs_pubmed_for_multiple_factors啟發和延伸，目的是為變成通用版
有時候get_pubmed_ids會出現warning: error in erl: 無法開啟...連結
20230215 已新增 return retrieve PMIDs
20230215 已解決 "ERBB2 AND Lung cancer AND drug" >> Note that only 0 PubMed IDs were retrieved (1170 were expected).
}
