% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcipubmedtool.R
\name{return_value_for_each_sentance}
\alias{return_value_for_each_sentance}
\title{輸入句子和關鍵字，回傳確實存在句子中的關鍵字
預設忽略英文字母大小寫}
\usage{
return_value_for_each_sentance(sentences, keywords)
}
\arguments{
\item{sentences}{輸入句子}

\item{keywords}{輸入關鍵字}
}
\value{
回傳確實存在於句子中的關鍵字
}
\description{
輸入句子和關鍵字，回傳確實存在句子中的關鍵字
預設忽略英文字母大小寫
}
