% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcipubmedtool.R
\name{run_easypubmed_shiny}
\alias{run_easypubmed_shiny}
\title{R search PMIDs}
\usage{
run_easypubmed_shiny(host_, port_)
}
\arguments{
\item{host_}{input number}

\item{port_}{input port}
}
\value{
run shiny app
}
\description{
R search PMIDs
}
