% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcipubmedtool.R
\name{run_paper_download}
\alias{run_paper_download}
\title{R download PDF}
\usage{
run_paper_download(host_, port_)
}
\arguments{
\item{host_}{input number}

\item{port_}{input port}
}
\value{
r shiny app for paper download
}
\description{
R download PDF
}
