% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcipubmedtool.R
\name{SNPs_pubmed_for_multiple_factors}
\alias{SNPs_pubmed_for_multiple_factors}
\title{Pubmed文獻搜尋爬蟲: 兩組關鍵子彼此排列組合進行搜尋}
\usage{
SNPs_pubmed_for_multiple_factors(path, SNPs, factors, filename)
}
\arguments{
\item{path}{想要輸出到哪個路徑資料夾儲存}

\item{SNPs}{主要關鍵字/SNP清單}

\item{factors}{次要關鍵字}

\item{filename}{依照不同搜尋結果存放在不同資料夾}
}
\value{
包含PMID的CSV file
}
\description{
針對G2 384突變位點所設計，可設定不同關鍵字組合，搜尋Pubmed，並把搜尋結果彙整成表格
SNPs為主要關鍵字, factors為次要關鍵字。
範例說明: 例如不同SNPs v.s. 不同人種，會有多個排列組合 \link{SNP} AND \link{Country}
ouput file: Yes (.csv)
}
