% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcipubmedtool.R
\name{SNPs_pubmed_for_paired_disease_and_SNP}
\alias{SNPs_pubmed_for_paired_disease_and_SNP}
\title{Pubmed文獻搜尋爬蟲: 已經預設配對好的關鍵字組合}
\usage{
SNPs_pubmed_for_paired_disease_and_SNP(
  path,
  disease_col,
  snp_col,
  data,
  filename
)
}
\arguments{
\item{path}{想要輸出到哪個路徑資料夾儲存}

\item{disease_col}{欄位名稱:第一組預設好的關鍵字組}

\item{snp_col}{欄位名稱:第二組預設好的關鍵字組}

\item{data}{包含兩組預設關鍵字組的完整檔案清單}

\item{filename}{輸出的檔案名稱}
}
\value{
CSV file
}
\description{
針對G2 384突變位點所設計，可設定不同關鍵字組合，搜尋Pubmed，並把搜尋結果彙整成表格
SNPs為主要關鍵字, factors為次要關鍵字。
範例說明: 例如不同SNPs v.s. 不同人種，會有多個排列組合 \link{SNP} AND \link{Country}
ouput file: Yes (.csv)
}
