% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcipubmedtool.R
\name{text_mining_pubmed_pdf}
\alias{text_mining_pubmed_pdf}
\title{輸入句子和關鍵字，回傳確實存在句子中的關鍵字
預設忽略英文字母大小寫
針對單個pmid，而一組rsID+Disease有多個pmids; 一個pmid有多個句子
輸入單一個SNP}
\usage{
text_mining_pubmed_pdf(path_, pmid_list, rsID, SNP, race, disease, odds)
}
\arguments{
\item{pmid_list}{用來搜尋資料夾中檔名為PMID的PDF檔}

\item{rsID}{第一關鍵字}

\item{SNP}{第二關鍵字}

\item{race}{第三關鍵字}

\item{disease}{第四關鍵字}

\item{odds}{第五關鍵字}

\item{path}{檔案存放路徑}
}
\value{
輸出大表格，包含rsID對應疾病
}
\description{
輸入句子和關鍵字，回傳確實存在句子中的關鍵字
預設忽略英文字母大小寫
針對單個pmid，而一組rsID+Disease有多個pmids; 一個pmid有多個句子
輸入單一個SNP
}
